requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.1
pymongo==4.6.1
fastapi==0.110.0
uvicorn==0.29.0